fix borrow solana value
